@@ -58,66 +58,65 @@
       "ts-node/register"
     ],
     "reporter": [
-      "lcov"
+      "text-lcov"
     ],
     "sourceMap": true,
     "instrument": true
   },
   "dependencies": {
-    "@ledgerhq/hw-transport": "4.24.0",
-    "@ledgerhq/hw-transport-node-hid": "4.24.0",
-    "aws-sdk": "^2.533.0",
+    "@ledgerhq/hw-transport": "5.11.0",
+    "@ledgerhq/hw-transport-node-hid": "5.11.0",
     "babel-runtime": "6.26.0",
-    "base-n": "git+https://github.com/yourcodesucks/base-n.git",
-    "basil-tezos-ledger": "0.0.1",
-    "big-integer": "^1.6.47",
-    "bip39": "^3.0.2",
-    "blakejs": "^1.1.0",
-    "bs58check": "^2.1.2",
-    "https": "^1.0.0",
-    "libsodium-wrappers-sumo": "0.7.5",
-    "loglevel": "^1.6.6",
+    "big-integer": "1.6.48",
+    "bip32-path": "0.4.2",
+    "bip39": "3.0.2",
+    "blakejs": "1.1.0",
+    "bs58check": "2.1.2",
+    "jsonpath-plus": "3.0.0",
+    "libsodium-wrappers-sumo": "0.7.6",
+    "loglevel": "1.6.8",
     "moo": "0.5.0",
-    "nearley": "2.18.0",
-    "node-fetch": "^2.6.0",
-    "zxcvbn": "^4.4.2"
+    "nearley": "2.19.1",
+    "node-fetch": "2.6.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.2",
-    "@types/chai-as-promised": "^7.1.2",
-    "@types/mocha": "^5.2.7",
-    "@types/nock": "^11.1.0",
-    "@types/node": "^12.7.5",
-    "@types/node-fetch": "^2.5.0",
-    "@types/zxcvbn": "^4.4.0",
-    "@typescript-eslint/eslint-plugin": "^2.2.0",
-    "@typescript-eslint/parser": "^2.2.0",
-    "awesome-typescript-loader": "^5.0.0",
-    "chai": "^4.1.2",
-    "chai-as-promised": "^7.1.1",
-    "copyfiles": "^2.1.1",
-    "coveralls": "^3.0.6",
-    "eslint": "^6.3.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-typescript": "^0.14.0",
-    "glob": "7.1.4",
-    "mocha": "^6.2.0",
-    "nock": "^11.3.3",
-    "nyc": "^14.1.1",
-    "request-promise": "^4.2.4",
-    "source-map-loader": "^0.2.4",
-    "source-map-support": "^0.5.13",
-    "terser-webpack-plugin": "^1.4.1",
-    "ts-loader": "^6.1.0",
-    "ts-node": "^8.3.0",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "typedoc": "^0.15.0",
-    "typedoc-plugin-markdown": "^2.1.1",
-    "typescript": "^3.5.3",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.8",
-    "webpack-node-externals": "^1.7.2"
+    "@types/chai": "4.2.11",
+    "@types/chai-as-promised": "7.1.2",
+    "@types/mocha": "7.0.2",
+    "@types/nock": "11.1.0",
+    "@types/node": "13.9.1",
+    "@types/node-fetch": "2.5.6",
+    "@types/zxcvbn": "4.4.0",
+    "@typescript-eslint/eslint-plugin": "2.29.0",
+    "@typescript-eslint/parser": "2.29.0",
+    "awesome-typescript-loader": "5.2.1",
+    "chai": "4.2.0",
+    "chai-as-promised": "7.1.1",
+    "copyfiles": "2.2.0",
+    "coveralls": "3.0.12",
+    "eslint": "6.8.0",
+    "eslint-plugin-import": "2.20.2",
+    "glob": "7.1.6",
+    "mocha": "7.1.1",
+    "nock": "12.0.3",
+    "nyc": "15.0.1",
+    "request-promise": "4.2.5",
+    "source-map-loader": "0.2.4",
+    "source-map-support": "0.5.18",
+    "terser-webpack-plugin": "2.3.5",
+    "ts-loader": "6.2.1",
+    "ts-node": "8.9.0",
+    "tsconfig-paths-webpack-plugin": "3.2.0",
+    "typedoc": "0.17.4",
+    "typedoc-plugin-markdown": "2.2.17",
+    "typescript": "3.8.3",
+    "webpack": "4.43.0",
+    "webpack-cli": "3.3.11",
+    "webpack-node-externals": "1.7.2"
+  },
+  "engines": {
+    "node": ">=12.16.3",
+    "npm": ">=6.14.4"
   },
   "eslintConfig": {
     "parser": "@typescript-eslint/parser",
