@@ -11,12 +11,12 @@ import FetchSelector from '../../../src/utils/FetchSelector';
 FetchSelector.setFetch(fetch);
 
 import { TezosConseilClient } from '../../../src/reporting/tezos/TezosConseilClient';
-import { TezosProtocolHelper } from '../../../src/chain/tezos/TezosProtocolHelper';
+import { BabylonDelegationHelper } from '../../../src/chain/tezos/contracts/BabylonDelegationHelper';
 import { tezosServer, conseilServer, keys, bakerAddress } from '../../TestAssets';
 
-describe('TezosProtocolHelper integration test suite', () => {
+describe('BabylonDelegationHelper integration test suite', () => {
     it('Deploy manager.tz P005 "upgrade" contract', async () => {
-        const nodeResult = await TezosProtocolHelper.deployManagerContract(tezosServer, keys, bakerAddress, 20000, 500000);
+        const nodeResult = await BabylonDelegationHelper.deployManagerContract(tezosServer, keys, bakerAddress, 20000, 500000);
         expect(nodeResult["operationGroupID"]).to.exist;
 
         const groupid = nodeResult["operationGroupID"].replace(/\"/g, '').replace(/\n/, '');
@@ -24,7 +24,7 @@ describe('TezosProtocolHelper integration test suite', () => {
     });
 
     it('Set a delegate on the "upgrade" contract', async () => {
-        const nodeResult = await TezosProtocolHelper.setDelegate(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 'tz3Q67aMz7gSMiQRcW729sXSfuMtkyAHYfqc', 20000);
+        const nodeResult = await BabylonDelegationHelper.setDelegate(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 'tz3Q67aMz7gSMiQRcW729sXSfuMtkyAHYfqc', 20000);
         expect(nodeResult["operationGroupID"]).to.exist;
 
         const groupid = nodeResult["operationGroupID"].replace(/\"/g, '').replace(/\n/, '');
@@ -32,7 +32,7 @@ describe('TezosProtocolHelper integration test suite', () => {
     });
 
     it('Clear the delegate on the "upgrade" contract', async () => {
-        const nodeResult = await TezosProtocolHelper.unSetDelegate(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 20000);
+        const nodeResult = await BabylonDelegationHelper.unSetDelegate(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 20000);
         expect(nodeResult["operationGroupID"]).to.exist;
 
         const groupid = nodeResult["operationGroupID"].replace(/\"/g, '').replace(/\n/, '');
@@ -40,7 +40,7 @@ describe('TezosProtocolHelper integration test suite', () => {
     });
 
     it('Return delegated funds to the "manager" account', async () => {
-        const nodeResult = await TezosProtocolHelper.withdrawDelegatedFunds(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 20000, 250000);
+        const nodeResult = await BabylonDelegationHelper.withdrawDelegatedFunds(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 20000, 250000);
         expect(nodeResult["operationGroupID"]).to.exist;
 
         const groupid = nodeResult["operationGroupID"].replace(/\"/g, '').replace(/\n/, '');
@@ -48,7 +48,7 @@ describe('TezosProtocolHelper integration test suite', () => {
     });
 
     it('Deposit additional funds into the "upgrade" contract', async () => {
-        const nodeResult = await TezosProtocolHelper.depositDelegatedFunds(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 20000, 250000)
+        const nodeResult = await BabylonDelegationHelper.depositDelegatedFunds(tezosServer, keys, 'KT1PxkYcmUqjM6Giqu8WKhPyMB5mCVkvWrXg', 20000, 250000)
         expect(nodeResult["operationGroupID"]).to.exist;
 
         const groupid = nodeResult["operationGroupID"].replace(/\"/g, '').replace(/\n/, '');
