@@ -10,6 +10,7 @@ import FetchSelector from '../../../src/utils/FetchSelector';
 FetchSelector.setFetch(fetch);
 
 import { TezosContractIntrospector } from '../../../src/chain/tezos/TezosContractIntrospector';
+import { TezosLanguageUtil } from '../../../src/chain/tezos/TezosLanguageUtil'
 import { EntryPoint } from '../../../src/types/tezos/ContractIntrospectionTypes';
 import { ConseilQueryBuilder } from '../../../src/reporting/ConseilQueryBuilder';
 import { ConseilOperator, ConseilFunction, ConseilSortDirection } from '../../../src/types/conseil/QueryTypes';
@@ -17,45 +18,49 @@ import { ConseilDataClient } from '../../../src/reporting/ConseilDataClient';
 import { conseilServer } from '../../TestAssets';
 
 describe('TezosContractIntrospector integration test suite', () => {
-    it('Generate entry points for Tezos Baker Registry (Alphanet)', async () => {
-        const result: EntryPoint[] = await TezosContractIntrospector.generateEntryPointsFromAddress(conseilServer, conseilServer.network, 'KT1NpCh6tNQDmbmAVbGLxwRBx8jJD4rEFnmC');
-
-        expect(result[0].name).to.equal('%_Liq_entry_updateName');
-        expect(result[0].parameters.length).to.equal(1);
-        expect(result[0].generateParameter('"param1"')).to.equal('(Left "param1")');
-
-        expect(result[1].name).to.equal('%_Liq_entry_updatePaymentAddress');
-        expect(result[1].parameters.length).to.equal(1);
-        expect(result[1].generateParameter('"param1"')).to.equal('(Right (Left "param1"))');
-
-        expect(result[2].name).to.equal('%_Liq_entry_updateTerms');
-        expect(result[2].parameters.length).to.equal(3);
-        expect(result[2].generateParameter(1, 2, 3)).to.equal('(Right (Right (Left (Pair 1 (Pair 2 3)))))');
-
-        expect(result[3].name).to.equal('%_Liq_entry_deleteRegistration');
-        expect(result[3].parameters.length).to.equal(1);
-        expect(result[3].generateParameter('Unit')).to.equal('(Right (Right (Right Unit)))');
-    });
-
     it('Process on-chain contracts', async () => {
+        const cap = 500;
         let contractQuery = ConseilQueryBuilder.blankQuery();
         contractQuery = ConseilQueryBuilder.addFields(contractQuery, 'account_id', 'script');
         contractQuery = ConseilQueryBuilder.addPredicate(contractQuery, 'account_id', ConseilOperator.STARTSWITH, ['KT1']);
         contractQuery = ConseilQueryBuilder.addPredicate(contractQuery, 'script', ConseilOperator.ISNULL, [], true);
         contractQuery = ConseilQueryBuilder.addAggregationFunction(contractQuery, 'account_id', ConseilFunction.count);
         contractQuery = ConseilQueryBuilder.addOrdering(contractQuery, 'count_account_id', ConseilSortDirection.DESC)
-        contractQuery = ConseilQueryBuilder.setLimit(contractQuery, 100);
+        contractQuery = ConseilQueryBuilder.setLimit(contractQuery, cap);
 
         const contractList = await ConseilDataClient.executeEntityQuery(conseilServer, 'tezos', conseilServer.network, 'accounts', contractQuery);
 
+        const total = contractList.length;
+        let skipped = 0;
         contractList.forEach(r => {
-            if (r['script'].startsWith('Unparsable code:')) { return; } // accounting for invalid Conseil results
+            if (r['script'].startsWith('Unparsable code:')) { skipped++; return; } // accounting for invalid Conseil results
 
             try {
                 let p = TezosContractIntrospector.generateEntryPointsFromCode(r['script']);
+                console.log(`processed ${TezosLanguageUtil.preProcessMichelsonScript(r['script'])[0]}`);
+                if (p.length === 0) { 
+                    console.log('no entry points found');
+                } else {
+                    console.log(`entry point${p.length === 1 ? '' : 's' }:`);
+                    p.forEach(e => { console.log(`  ${formatEntryPoint(e)}`); });
+                }
             } catch (error) {
                 console.log(`error ${error}\n----\n${r['script']}\n----\n`);
             }
         });
+
+        console.log(`found ${total} contracts${skipped ? ', skipped ' + skipped : ''}${cap === total ? ', more may be available' : ''}`);
     });
 });
+
+function formatEntryPoint(entryPoint: EntryPoint){
+    let f = entryPoint.name ? `${entryPoint.name}: ` : '' ;
+
+    let args = entryPoint.structure;
+    for (let i = 0 ; i < entryPoint.parameters.length; i++) {
+        let param = entryPoint.parameters[i];
+        args = args.replace('$PARAM', `${param.type}${param.name ? ' %' + param.name : ''}`);
+    }
+
+    return f + args;
+}
