@@ -1,4 +1,4 @@
-// Generated automatically by nearley, version 2.18.0
+// Generated automatically by nearley, version 2.19.1
 // http://github.com/Hardmath123/nearley
 // Bypasses TS6133. Allow declared but unused functions.
 // @ts-ignore
@@ -24,8 +24,8 @@ declare var data: any;
         parameter: 'parameter',
         or: 'or',
         pair: 'pair',
-        data: ['bytes', 'int', 'nat', 'bool', 'string', 'timestamp', 'signature', 'key', 'key_hash', 'mutez', 'address', 'unit', 'operation'],
-        singleArgData: ['option', 'list', 'contract'],
+        data: ['bytes', 'int', 'nat', 'bool', 'string', 'timestamp', 'signature', 'key', 'key_hash', 'mutez', 'address', 'unit', 'operation', 'chain_id'],
+        singleArgData: ['option', 'list', 'contract', 'set'],
         doubleArgData: ['lambda', 'map', 'big_map'],
         semicolon: ';'
     });
@@ -46,13 +46,14 @@ declare var data: any;
         const rightEntryPoints: EntryPoint[] = d[8];
         const branchedEntryPoints: EntryPoint[] = [];
 
-        console.log(`branchOrWithTwoAnnot found ${annotA}, ${annotB}`);
         for (const leftEntryPoint of leftEntryPoints) {
             const branchedEntryPoint: EntryPoint = {
                 name: leftEntryPoint.name, // TODO
                 parameters: leftEntryPoint.parameters,
                 structure: '(Left ' + leftEntryPoint.structure + ')',
-                generateParameter: leftEntryPoint.generateParameter
+                generateInvocationString: leftEntryPoint.generateInvocationString,
+                generateInvocationPair: leftEntryPoint.generateInvocationPair,
+                generateSampleInvocation: leftEntryPoint.generateSampleInvocation
             }
             branchedEntryPoints.push(branchedEntryPoint);
         }
@@ -62,7 +63,9 @@ declare var data: any;
                 name: rightEntryPoint.name, // TODO
                 parameters: rightEntryPoint.parameters,
                 structure: '(Right ' + rightEntryPoint.structure + ')',
-                generateParameter: rightEntryPoint.generateParameter
+                generateInvocationString: rightEntryPoint.generateInvocationString,
+                generateInvocationPair: rightEntryPoint.generateInvocationPair,
+                generateSampleInvocation: rightEntryPoint.generateSampleInvocation
             }
             branchedEntryPoints.push(branchedEntryPoint);
         }
@@ -76,13 +79,14 @@ declare var data: any;
         const rightEntryPoints: EntryPoint[] = d[6];
         const branchedEntryPoints: EntryPoint[] = [];
 
-        console.log(`branchOrWithAnnot found ${annot}`);
         for (const leftEntryPoint of leftEntryPoints) {
             const branchedEntryPoint: EntryPoint = {
                 name: `${annot}.${leftEntryPoint.name}`, // TODO
                 parameters: leftEntryPoint.parameters,
                 structure: '(Left ' + leftEntryPoint.structure + ')',
-                generateParameter: leftEntryPoint.generateParameter
+                generateInvocationString: leftEntryPoint.generateInvocationString,
+                generateInvocationPair: leftEntryPoint.generateInvocationPair,
+                generateSampleInvocation: leftEntryPoint.generateSampleInvocation
             }
             branchedEntryPoints.push(branchedEntryPoint);
         }
@@ -92,7 +96,9 @@ declare var data: any;
                 name: `${annot}.${rightEntryPoint.name}`, // TODO
                 parameters: rightEntryPoint.parameters,
                 structure: '(Right ' + rightEntryPoint.structure + ')',
-                generateParameter: rightEntryPoint.generateParameter
+                generateInvocationString: rightEntryPoint.generateInvocationString,
+                generateInvocationPair: rightEntryPoint.generateInvocationPair,
+                generateSampleInvocation: rightEntryPoint.generateSampleInvocation
             }
             branchedEntryPoints.push(branchedEntryPoint);
         }
@@ -106,22 +112,17 @@ declare var data: any;
         const branchedEntryPoints: EntryPoint[] = [];
 
         for (const leftEntryPoint of leftEntryPoints) {
-            const branchedEntryPoint: EntryPoint = {
-                name: leftEntryPoint.name,
-                parameters: leftEntryPoint.parameters,
-                structure: '(Left ' + leftEntryPoint.structure + ')',
-                generateParameter: leftEntryPoint.generateParameter
+            if (leftEntryPoint.parameters.length === 1 && leftEntryPoint.parameters[0].name === leftEntryPoint.name) {
+                leftEntryPoint.parameters[0].name = undefined;
             }
-            branchedEntryPoints.push(branchedEntryPoint);
+            branchedEntryPoints.push({...leftEntryPoint, structure: `(Left ${leftEntryPoint.structure})`});
         }
+
         for (const rightEntryPoint of rightEntryPoints) {
-            const branchedEntryPoint: EntryPoint = {
-                name: rightEntryPoint.name,
-                parameters: rightEntryPoint.parameters,
-                structure: '(Right ' + rightEntryPoint.structure + ')',
-                generateParameter: rightEntryPoint.generateParameter
+            if (rightEntryPoint.parameters.length === 1 && rightEntryPoint.parameters[0].name === rightEntryPoint.name) {
+                rightEntryPoint.parameters[0].name = undefined;
             }
-            branchedEntryPoints.push(branchedEntryPoint);
+            branchedEntryPoints.push({...rightEntryPoint, structure: `(Right ${rightEntryPoint.structure})`});
         }
 
         return branchedEntryPoints;
@@ -136,14 +137,15 @@ declare var data: any;
         const secondEntryPoints: EntryPoint[] = d[8];
         const pairedEntryPoints: EntryPoint[] = [];
 
-        console.log(`mergePairWithTwoAnnot found ${annotA}, ${annotB}`);
         for (const firstEntryPoint of firstEntryPoints) {
             for (const secondEntryPoint of secondEntryPoints) {
                 const pairedEntryPoint: EntryPoint = {
-                    name: annotA.toString(), // TODO
+                    name: getTypeAnnotation(annotA.toString(), annotB.toString()),
                     parameters: firstEntryPoint.parameters.concat(secondEntryPoint.parameters),
                     structure: `(Pair ${firstEntryPoint.structure} ${secondEntryPoint.structure})`,
-                    generateParameter: firstEntryPoint.generateParameter
+                    generateInvocationString: firstEntryPoint.generateInvocationString,
+                    generateInvocationPair: firstEntryPoint.generateInvocationPair,
+                    generateSampleInvocation: firstEntryPoint.generateSampleInvocation
                 }
                 pairedEntryPoints.push(pairedEntryPoint);
             }
@@ -158,14 +160,16 @@ declare var data: any;
         const secondEntryPoints: EntryPoint[] = d[6];
         const pairedEntryPoints: EntryPoint[] = [];
 
-        console.log(`mergePairWithAnnot found ${annot}`);
         for (const firstEntryPoint of firstEntryPoints) {
             for (const secondEntryPoint of secondEntryPoints) {
+                const name = getTypeAnnotation(annot.toString()) || getFieldAnnotation(annot.toString());
                 const pairedEntryPoint: EntryPoint = {
-                    name: annot.toString(), // TODO
+                    name: name || undefined,
                     parameters: firstEntryPoint.parameters.concat(secondEntryPoint.parameters),
                     structure: `(Pair ${firstEntryPoint.structure} ${secondEntryPoint.structure})`,
-                    generateParameter: firstEntryPoint.generateParameter
+                    generateInvocationString: firstEntryPoint.generateInvocationString,
+                    generateInvocationPair: firstEntryPoint.generateInvocationPair,
+                    generateSampleInvocation: firstEntryPoint.generateSampleInvocation // TODO
                 }
                 pairedEntryPoints.push(pairedEntryPoint);
             }
@@ -181,17 +185,13 @@ declare var data: any;
 
         for (const firstEntryPoint of firstEntryPoints) {
             for (const secondEntryPoint of secondEntryPoints) {
-                let pairedEntryPointName: string | undefined = undefined;
-                if (firstEntryPoint.name != undefined) {
-                    pairedEntryPointName = firstEntryPoint.name;
-                } else if (secondEntryPoint.name != undefined) {
-                    pairedEntryPointName = secondEntryPoint.name;
-                }
                 const pairedEntryPoint: EntryPoint = {
-                    name: pairedEntryPointName,
+                    name: undefined,
                     parameters: firstEntryPoint.parameters.concat(secondEntryPoint.parameters),
                     structure: `(Pair ${firstEntryPoint.structure} ${secondEntryPoint.structure})`,
-                    generateParameter: firstEntryPoint.generateParameter
+                    generateInvocationString: firstEntryPoint.generateInvocationString,
+                    generateInvocationPair: firstEntryPoint.generateInvocationPair,
+                    generateSampleInvocation: firstEntryPoint.generateSampleInvocation // TODO
                 }
                 pairedEntryPoints.push(pairedEntryPoint);
             }
@@ -208,8 +208,9 @@ declare var data: any;
         const annotB: string = d[4].toString();
         const entryPoints: EntryPoint[] = d[6];
 
-        console.log(`recordSingleArgDataWithTwoAnnot found ${annotA}, ${annotB}`);
-        //entryPoints[0].parameters[0].name = annot;
+        entryPoints[0].name = getFieldAnnotation(annotA, annotB);
+        entryPoints[0].parameters[0].constituentType = entryPoints[0].parameters[0].type;
+        if (singleArgData === 'option') { entryPoints[0].parameters[0].optional = true; }
         entryPoints[0].parameters[0].type = `${singleArgData} (${entryPoints[0].parameters[0].type})`;
         entryPoints[0].structure = `(${entryPoints[0].structure})`;
 
@@ -221,8 +222,9 @@ declare var data: any;
         const annot: string = d[2].toString();
         const entryPoints: EntryPoint[] = d[4];
 
-        console.log(`recordSingleArgDataWithAnnot found ${annot}`);
-        entryPoints[0].parameters[0].name = annot;
+        entryPoints[0].name = getFieldAnnotation(annot);
+        entryPoints[0].parameters[0].constituentType = entryPoints[0].parameters[0].type;
+        if (singleArgData === 'option') { entryPoints[0].parameters[0].optional = true; }
         entryPoints[0].parameters[0].type = `${singleArgData} (${entryPoints[0].parameters[0].type})`;
         entryPoints[0].structure = `(${entryPoints[0].structure})`;
 
@@ -233,6 +235,8 @@ declare var data: any;
         const singleArgData: string = d[0].toString();
         const entryPoints: EntryPoint[] = d[2];
 
+        entryPoints[0].parameters[0].constituentType = entryPoints[0].parameters[0].type;
+        if (singleArgData === 'option') { entryPoints[0].parameters[0].optional = true; }
         entryPoints[0].parameters[0].type = `${singleArgData} (${entryPoints[0].parameters[0].type})`;
         entryPoints[0].structure = `(${entryPoints[0].structure})`;
 
@@ -248,8 +252,7 @@ declare var data: any;
         const firstEntryPoints: EntryPoint[] = d[6];
         const secondEntryPoints: EntryPoint[] = d[8];
 
-        //firstEntryPoints[0].parameters[0].name = annot;
-        console.log(`recordDoubleArgDataWithTwoAnnot found ${annotA}, ${annotB}`);
+        firstEntryPoints[0].name = getFieldAnnotation(annotA, annotB);
         firstEntryPoints[0].parameters[0].type = `${doubleArgData} (${firstEntryPoints[0].parameters[0].type}) (${secondEntryPoints[0].parameters[0].type})`;
         firstEntryPoints[0].structure = `(${firstEntryPoints[0].structure})`;
 
@@ -262,8 +265,7 @@ declare var data: any;
         const firstEntryPoints: EntryPoint[] = d[4];
         const secondEntryPoints: EntryPoint[] = d[6];
 
-        console.log(`recordDoubleArgDataWithAnnot found ${annot}`);
-        firstEntryPoints[0].parameters[0].name = annot;
+        firstEntryPoints[0].name = getFieldAnnotation(annot);
         firstEntryPoints[0].parameters[0].type = `${doubleArgData} (${firstEntryPoints[0].parameters[0].type}) (${secondEntryPoints[0].parameters[0].type})`;
         firstEntryPoints[0].structure = `(${firstEntryPoints[0].structure})`;
 
@@ -289,6 +291,7 @@ declare var data: any;
 
         if (d.length >= 3) {
             const annot = d[2].toString();
+
             if (annot.charAt(0) === '%') {
                 entryPointName = formatFieldAnnotation(annot);
             } else {
@@ -298,6 +301,7 @@ declare var data: any;
 
         if (d.length === 5) {
             const anotherAnnot = d[4].toString();
+
             if (anotherAnnot.startsWith('%') && entryPointName === undefined) {
                 entryPointName = formatFieldAnnotation(anotherAnnot);
             }
@@ -307,7 +311,7 @@ declare var data: any;
         }
 
         const parameter: Parameter = {
-            name: parameterName,
+            name: parameterName || entryPointName,
             type: d[0].toString()
         };
 
@@ -315,13 +319,65 @@ declare var data: any;
             name: entryPointName,
             parameters: [parameter],
             structure: '$PARAM',
-            generateParameter(...vars: any[]): string {
+            generateInvocationString(...vars: any[]): string { // TODO: check double-quotes on string-like parameters; support for map type
                 if (this.parameters.length !== vars.length) { throw new Error(`Incorrect number of parameters provided; expected ${this.parameters.length}, got ${vars.length}`); }
                 let invocationParameter: string = this.structure;
                 for (let i = 0 ; i < this.parameters.length; i++) {
-                    invocationParameter = invocationParameter.replace('$PARAM', vars[i]);
+                    let val = vars[i];
+
+                    if (this.parameters[i].type === 'unit') { val = 'Unit'; }
+
+                    if (this.parameters[i].type.startsWith('list')) {
+                        if (!Array.isArray(val)) { throw new Error(`${JSON.stringify(this.parameters[i])} requires an array value`); }
+                        val = `{${val.join('; ')}}`;
+                    }
+
+                    if (this.parameters[i].optional && vars[i]) {
+                        val = `Some ${val}`;
+                    } else if (this.parameters[i].optional && !vars[i]) { 
+                        val = 'None';
+                    }
+
+                    invocationParameter = invocationParameter.replace('$PARAM', val);
                 }
                 return invocationParameter;
+            },
+            generateInvocationPair(...vars: any[]): any {
+                let param = this.generateInvocationString(...vars);
+
+                while (param.startsWith('(Left ') || param.startsWith('(Right ')) {
+                    if (param.startsWith('(Left ')) {
+                        param = param.slice(6, -1);
+                    }
+                    if (param.startsWith('(Right ')) {
+                        param = param.slice(7, -1);
+                    }
+                }
+                return { entrypoint: this.name, parameters: param };
+            },
+            generateSampleInvocation(): string {
+                const params = this.parameters.map(p => {
+                    switch (p.type) {
+                        case 'string': { return '"Tacos"'; }
+                        case 'int': { return -1; }
+                        case 'nat': { return 99; }
+                        case 'address': { return '"KT1EGbAxguaWQFkV3Egb2Z1r933MWuEYyrJS"'; }
+                        case 'key_hash': { return '"tz1SQnJaocpquTztY3zMgydTPoQBBQrDGonJ"'; }
+                        case 'timestamp': { return `"${(new Date()).toISOString()}"`}
+                        case 'mutez': { return 500000; }
+                        case 'unit': { return 'Unit'; }
+                        case 'bytes':
+                        case 'bool':
+                        case 'signature':
+                        case 'key':
+                        case 'operation':
+                        case 'chain_id':
+                        default: { return p.type; }
+                    }
+                });
+
+                return this.generateInvocationString(...params);
+                
             }
         };
 
@@ -330,57 +386,57 @@ declare var data: any;
 
     const getFieldAnnotation = (...annot: string[]) => {
         const fa = annot.find(a => a.startsWith('%'));
-        if (!!fa) {
-            return formatFieldAnnotation(fa);
-        }
 
-        return undefined;
+        return !!fa ? formatFieldAnnotation(fa): undefined;
     }
 
     const getTypeAnnotation = (...annot: string[]) => {
         const ta = annot.find(a => a.startsWith(':'));
-        if (!!ta) {
-            return formatTypeAnnotation(ta);
-        }
 
-        return undefined;
+        return !!ta ? formatTypeAnnotation(ta): undefined;
     }
 
     const formatFieldAnnotation = (annot: string) => {
         if (!annot.startsWith('%')) { throw new Error(`${annot} must start with '%'`); }
 
-        let name = annot.replace(/^%_Liq_entry_/, '').replace('%', '');
-        return name.charAt(0).toUpperCase() + name.slice(1);
+        return annot.replace(/^%_Liq_entry_/, '').replace('%', '');
     }
 
     const formatTypeAnnotation = (annot: string) => {
         if (!annot.startsWith(':')) { throw new Error(`${annot} must start with ':'`); }
 
-        let name = annot.replace(':', '');
-        return name.charAt(0).toUpperCase() + name.slice(1);
+        return annot.replace(':', '');
     }
 
-export interface Token { value: any; [key: string]: any };
-
-export interface Lexer {
-  reset: (chunk: string, info: any) => void;
-  next: () => Token | undefined;
-  save: () => any;
-  formatError: (token: Token) => string;
-  has: (tokenType: string) => boolean
+interface NearleyToken {  value: any;
+  [key: string]: any;
 };
 
-export interface NearleyRule {
+interface NearleyLexer {
+  reset: (chunk: string, info: any) => void;
+  next: () => NearleyToken | undefined;
+  save: () => any;
+  formatError: (token: NearleyToken) => string;
+  has: (tokenType: string) => boolean;
+};
+
+interface NearleyRule {
   name: string;
   symbols: NearleySymbol[];
-  postprocess?: (d: any[], loc?: number, reject?: {}) => any
+  postprocess?: (d: any[], loc?: number, reject?: {}) => any;
 };
 
-export type NearleySymbol = string | { literal: any } | { test: (token: any) => boolean };
+type NearleySymbol = string | { literal: any } | { test: (token: any) => boolean };
 
-export var Lexer: Lexer | undefined = lexer;
+interface Grammar {
+  Lexer: NearleyLexer | undefined;
+  ParserRules: NearleyRule[];
+  ParserStart: string;
+};
 
-export var ParserRules: NearleyRule[] = [
+const grammar: Grammar = {
+  Lexer: lexer,
+  ParserRules: [
     {"name": "entry", "symbols": [(lexer.has("parameter") ? {type: "parameter"} : parameter), "__", "parameters", "_", (lexer.has("semicolon") ? {type: "semicolon"} : semicolon)], "postprocess": breakParameter},
     {"name": "parameters", "symbols": [(lexer.has("lparen") ? {type: "lparen"} : lparen), "_", "parameters", "_", (lexer.has("rparen") ? {type: "rparen"} : rparen)], "postprocess": stripParen},
     {"name": "parameters", "symbols": [(lexer.has("or") ? {type: "or"} : or), "_", (lexer.has("annot") ? {type: "annot"} : annot), "__", (lexer.has("annot") ? {type: "annot"} : annot), "__", "parameters", "__", "parameters"], "postprocess": branchOrWithTwoAnnot},
@@ -402,6 +458,8 @@ export var ParserRules: NearleyRule[] = [
     {"name": "_$ebnf$1", "symbols": ["_$ebnf$1", /[\s]/], "postprocess": (d) => d[0].concat([d[1]])},
     {"name": "_", "symbols": ["_$ebnf$1"]},
     {"name": "__", "symbols": [/[\s]/]}
-];
+  ],
+  ParserStart: "entry",
+};
 
-export var ParserStart: string = "entry";
+export default grammar;
