@@ -9,7 +9,14 @@ export * from './chain/tezos/TezosLanguageUtil';
 export * from './chain/tezos/TezosMessageUtil';
 export * from './chain/tezos/TezosNodeReader';
 export * from './chain/tezos/TezosNodeWriter';
-export * from './chain/tezos/TezosProtocolHelper';
+export * from './chain/tezos/contracts/BabylonDelegationHelper';
+export * from './chain/tezos/contracts/CryptonomicNameServiceHelper';
+export * from './chain/tezos/contracts/DexterTokenHelper';
+export * from './chain/tezos/contracts/MurbardMultisigHelper';
+export * from './chain/tezos/contracts/StakerDAOTokenHelper';
+export * from './chain/tezos/contracts/TCFBakerRegistryHelper';
+export * from './chain/tezos/contracts/Tzip7ReferenceTokenHelper';
+export * from './chain/tezos/contracts/TzbtcTokenHelper';
 
 export * from './identity/tezos/TezosWalletUtil';
 export * from './identity/tezos/TezosFileWallet';
