@@ -1,8 +1,8 @@
 import fetch from 'node-fetch';
 import * as log from 'loglevel';
 
-import DeviceSelector from './utils/DeviceSelector';
 import FetchSelector from './utils/FetchSelector';
+import DeviceSelector from './utils/DeviceSelector';
 import LogSelector from './utils/LoggerSelector';
 
 FetchSelector.setFetch(fetch);
@@ -16,13 +16,21 @@ export function setLogLevel(level: string) {
 export * from './chain/tezos/TezosContractIntrospector';
 export * from './chain/tezos/TezosLanguageUtil';
 export * from './chain/tezos/TezosMessageUtil';
-export * from './chain/tezos/TezosNodeReader';
-export * from './chain/tezos/TezosNodeWriter';
-export * from './chain/tezos/TezosProtocolHelper';
-export * from './identity/tezos/TezosFileWallet';
-export * from './identity/tezos/TezosWalletUtil';
+export * from "./chain/tezos/TezosNodeReader";
+export * from "./chain/tezos/TezosNodeWriter";
+export * from './chain/tezos/contracts/BabylonDelegationHelper';
+export * from './chain/tezos/contracts/CryptonomicNameServiceHelper';
+export * from './chain/tezos/contracts/DexterTokenHelper';
+export * from './chain/tezos/contracts/MurbardMultisigHelper';
+export * from './chain/tezos/contracts/StakerDAOTokenHelper';
+export * from './chain/tezos/contracts/TCFBakerRegistryHelper';
+export * from './chain/tezos/contracts/Tzip7ReferenceTokenHelper';
+export * from './chain/tezos/contracts/TzbtcTokenHelper';
 
-export * from './reporting/tezos/TezosConseilClient';
+export * from "./identity/tezos/TezosFileWallet";
+export * from "./identity/tezos/TezosWalletUtil";
+
+export * from "./reporting/tezos/TezosConseilClient";
 
 export * from './reporting/ConseilDataClient';
 export * from './reporting/ConseilMetadataClient';
@@ -35,6 +43,7 @@ export * from './types/tezos/TezosChainTypes';
 export * from './types/tezos/TezosP2PMessageTypes';
 export * from './types/tezos/TezosRPCResponseTypes';
 export * from './types/wallet/KeyStore';
+
 export * from './utils/CryptoUtils';
 
 let TezosLedgerWallet;
