@@ -20,7 +20,7 @@ export namespace ConseilDataClient {
      * @param query JSON object or text confirming to the Conseil query spec.
      */
     export async function executeEntityQuery(serverInfo: ConseilServerInfo, platform: string, network: string, entity: string, query: ConseilQuery): Promise<any[]> {
-        const url = `${serverInfo.url}/v2/data/${platform}/${network}/${entity}`
+        const url = `${serverInfo.url}/v2/data/${platform}/${network}/${entity}`;
         log.debug(`ConseilDataClient.executeEntityQuery request: ${url}, ${JSON.stringify(query)}`);
 
         return fetch(url, {
