@@ -1,11 +1,15 @@
 export interface Parameter {
-    name?: string;
     type: string;
+    name?: string;
+    optional?: boolean;
+    constituentType?: string;
 }
 
 export interface EntryPoint {
     name?: string;
     parameters: Parameter[];
     structure: string;
-    generateParameter(... vars: any[]): string;
+    generateInvocationString(...vars: any[]): string;
+    generateInvocationPair(...vars: any[]): { entrypoint: string, parameters: any };
+    generateSampleInvocation(): string;
 }
