@@ -123,7 +123,7 @@ export interface TezosBlockOperationContentMetadataOperationResult {
 export interface Contract {
     balance: string;
     delegate?: ContractDelegate;
-    script?: string;
+    script?: any;
     counter: string;
 }
 
