@@ -24,12 +24,14 @@ import {
 
 use(chaiAsPromised);
 
+const conseilServer = { url: 'http://conseil.server', apiKey: 'c0ffee', network: 'alphanet' };
+
 describe('TezosConseilClient tests', () => {
     it('TezosConseilClient.getBlockHead', mochaAsync(async () => {
         const nockedserver = nock('http://conseil.server');
         nockedserver.post('/v2/data/tezos/alphanet/blocks').reply(200, blockHead);
 
-        const result = await TezosConseilClient.getBlockHead({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet');
+        const result = await TezosConseilClient.getBlockHead(conseilServer, conseilServer.network);
 
         expect(result['level']).to.equal(173066);
     }));
@@ -38,19 +40,17 @@ describe('TezosConseilClient tests', () => {
         const nockedserver = nock('http://conseil.server');
         nockedserver.post('/v2/data/tezos/alphanet/blocks').reply(200, block);
 
-        const result = await TezosConseilClient.getBlock({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', 'BL5zoNBN17j2AcUrs8mqSKSMcEiuBKkd9RB6uZ6CgYE2Xyb2ybV');
+        const result = await TezosConseilClient.getBlock(conseilServer, conseilServer.network, 'BM9bKSFuZs7rE6Aa5p6Lx1QhqcUeJNf7kEiSEtqAxhz5C2a7mz9');
 
-        expect(result.length).to.equal(1);
-        expect(result[0]['predecessor']).to.equal('BLyxiXprmaDkCeZo3b9JHU4udjPiVUpuTR1eKXSxtJe9o8JMbiM');
+        expect(result['predecessor']).to.equal('BLyxiXprmaDkCeZo3b9JHU4udjPiVUpuTR1eKXSxtJe9o8JMbiM');
     }));
 
     it('TezosConseilClient.getBlockByLevel', mochaAsync(async () => {
         const nockedserver = nock('http://conseil.server');
         nockedserver.post('/v2/data/tezos/alphanet/blocks').reply(200, block);
 
-        const result = await TezosConseilClient.getBlockByLevel({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', 173137);
+        const result = await TezosConseilClient.getBlockByLevel(conseilServer, conseilServer.network, 173137);
 
-        expect(result.length).to.equal(1);
         expect(result[0]['predecessor']).to.equal('BLyxiXprmaDkCeZo3b9JHU4udjPiVUpuTR1eKXSxtJe9o8JMbiM');
     }));
 
@@ -58,20 +58,18 @@ describe('TezosConseilClient tests', () => {
         const nockedserver = nock('http://conseil.server');
         nockedserver.post('/v2/data/tezos/alphanet/accounts').reply(200, account);
 
-        const result = await TezosConseilClient.getAccount({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', 'tz1L8MjMxQJio8YmmfdFbbVSymbNx5uiX3XT');
+        const result = await TezosConseilClient.getAccount(conseilServer, conseilServer.network, 'tz1L8MjMxQJio8YmmfdFbbVSymbNx5uiX3XT');
 
-        expect(result.length).to.equal(1);
-        expect(result[0]['manager']).to.equal('tz1L8MjMxQJio8YmmfdFbbVSymbNx5uiX3XT');
+        expect(result['manager']).to.equal('tz1L8MjMxQJio8YmmfdFbbVSymbNx5uiX3XT');
     }));
 
     it('TezosConseilClient.getOperationGroup', mochaAsync(async () => {
         const nockedserver = nock('http://conseil.server');
         nockedserver.post('/v2/data/tezos/alphanet/operation_groups').reply(200, operationgroup);
 
-        const result = await TezosConseilClient.getOperationGroup({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', 'ooYPxHKwUgkXxbWVfZFuYnvjNuGtrAgRQEtNrbPXNiwTBbCFW9j');
+        const result = await TezosConseilClient.getOperationGroup(conseilServer, conseilServer.network, 'ooYPxHKwUgkXxbWVfZFuYnvjNuGtrAgRQEtNrbPXNiwTBbCFW9j');
 
-        expect(result.length).to.equal(1);
-        expect(result[0]['branch']).to.equal('BLA3ecN7jh6Vc7XmogePRsdAjjbueG9NHsb32ByL5nbAhMvcNxK');
+        expect(result['branch']).to.equal('BLA3ecN7jh6Vc7XmogePRsdAjjbueG9NHsb32ByL5nbAhMvcNxK');
     }));
 
     it('TezosConseilClient.getBlocks', mochaAsync(async () => {
@@ -79,7 +77,7 @@ describe('TezosConseilClient tests', () => {
         nockedserver.post('/v2/data/tezos/alphanet/blocks').reply(200, blocks);
 
         const query = ConseilQueryBuilder.setLimit(ConseilQueryBuilder.blankQuery(), 3);
-        const result = await TezosConseilClient.getBlocks({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getBlocks(conseilServer, conseilServer.network, query);
 
         expect(result.length).to.equal(3);
         expect(result[1]['predecessor']).to.equal('BLqVwN6edWdVjER62uoF9N3KcMXZc4or3j2NetsHQnzNJCUwBbR');
@@ -90,7 +88,7 @@ describe('TezosConseilClient tests', () => {
         nockedserver.post('/v2/data/tezos/alphanet/accounts').reply(200, accounts);
 
         const query = ConseilQueryBuilder.setLimit(ConseilQueryBuilder.blankQuery(), 2);
-        const result = await TezosConseilClient.getAccounts({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getAccounts(conseilServer, conseilServer.network, query);
 
         expect(result.length).to.equal(2);
         expect(result[1]['counter']).to.equal(23);
@@ -101,7 +99,7 @@ describe('TezosConseilClient tests', () => {
         nockedserver.post('/v2/data/tezos/alphanet/operation_groups').reply(200, operationgroups);
 
         const query = ConseilQueryBuilder.setLimit(ConseilQueryBuilder.blankQuery(), 3);
-        const result = await TezosConseilClient.getOperationGroups({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getOperationGroups(conseilServer, conseilServer.network, query);
 
         expect(result.length).to.equal(3);
         expect(result[1]['block_id']).to.equal('BLPcavx4rhmZqCot3aZ3TiHdjQ4Ad5Tzs8WPmwJEeuxPx7F2GvQ');
@@ -115,7 +113,7 @@ describe('TezosConseilClient tests', () => {
         query = ConseilQueryBuilder.addPredicate(query, 'kind', ConseilOperator.EQ, ['transaction'], false);
         query = ConseilQueryBuilder.addOrdering(query, 'block_level', ConseilSortDirection.DESC);
         query = ConseilQueryBuilder.setLimit(query, 5);
-        const result = await TezosConseilClient.getOperations({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getOperations(conseilServer, conseilServer.network, query);
 
         expect(result.length).to.equal(5);
         expect(result[1]['timestamp']).to.be.greaterThan(result[2]['timestamp']);
@@ -125,7 +123,7 @@ describe('TezosConseilClient tests', () => {
         const nockedserver = nock('http://conseil.server');
         nockedserver.post('/v2/data/tezos/alphanet/fees').reply(200, transactionfees);
 
-        const fees = await TezosConseilClient.getFeeStatistics({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', OperationKindType.Transaction);
+        const fees = await TezosConseilClient.getFeeStatistics(conseilServer, conseilServer.network, OperationKindType.Transaction);
 
         expect(fees[0]['low']).to.lessThan(fees[0]['medium']);
         expect(fees[0]['medium']).to.lessThan(fees[0]['high']);
@@ -136,7 +134,7 @@ describe('TezosConseilClient tests', () => {
         nockedserver.post('/v2/data/tezos/alphanet/proposals').reply(200, proposals);
 
         const query = ConseilQueryBuilder.setLimit(ConseilQueryBuilder.blankQuery(), 3);
-        const result = await TezosConseilClient.getProposals({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getProposals(conseilServer, conseilServer.network, query);
 
         expect(result.length).to.equal(0);
     }));
@@ -146,7 +144,7 @@ describe('TezosConseilClient tests', () => {
         nockedserver.post('/v2/data/tezos/alphanet/bakers').reply(200, bakers);
 
         const query = ConseilQueryBuilder.setLimit(ConseilQueryBuilder.blankQuery(), 3);
-        const result = await TezosConseilClient.getBakers({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getBakers(conseilServer, conseilServer.network, query);
 
         expect(result.length).to.equal(3);
         expect(result[1]['pkh']).to.equal('tz1hodJSw6uv7LqArLW86zKuUjkXiayJvqCf');
@@ -157,7 +155,7 @@ describe('TezosConseilClient tests', () => {
         nockedserver.post('/v2/data/tezos/alphanet/ballots').reply(200, ballots);
 
         const query = ConseilQueryBuilder.setLimit(ConseilQueryBuilder.blankQuery(), 3);
-        const result = await TezosConseilClient.getBallots({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getBallots(conseilServer, conseilServer.network, query);
 
         expect(result.length).to.equal(0);
     }));
@@ -166,7 +164,7 @@ describe('TezosConseilClient tests', () => {
         const nockedserver = nock('http://conseil.server');
         nockedserver.post('/v2/data/tezos/alphanet/blocks').reply(404, blockHead);
 
-        await expect(TezosConseilClient.getBlock({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', 'BL5zoNBN17j2AcUrs8mqSKSMcEiuBKkd9RB6uZ6CgYE2Xyb2ybV')).to.be.rejected;
+        await expect(TezosConseilClient.getBlock(conseilServer, conseilServer.network, 'BL5zoNBN17j2AcUrs8mqSKSMcEiuBKkd9RB6uZ6CgYE2Xyb2ybV')).to.be.rejected;
     });
 
     it('TezosConseilClient.getOperations as CSV', mochaAsync(async () => {
@@ -178,7 +176,7 @@ describe('TezosConseilClient tests', () => {
         query = ConseilQueryBuilder.addOrdering(query, 'block_level', ConseilSortDirection.DESC);
         query = ConseilQueryBuilder.setLimit(query, 5);
         query = ConseilQueryBuilder.setOutputType(query, ConseilOutput.csv);
-        const result = await TezosConseilClient.getOperations({ url: 'http://conseil.server', apiKey: 'c0ffee' }, 'alphanet', query);
+        const result = await TezosConseilClient.getOperations(conseilServer, conseilServer.network, query);
 
         expect(result.toString().split('\n').length).to.equal(6);
     }));
